requirements.txt
Django
eyed3
instagrapi
moviepy
pedalboard
Pillow
pydub
requests
requests_oauthlib
spotipy
tweepy
spotdl
libsass
django-compressor 
django-sass-processor
django-unused-media
colorgram.py
scdl
pytest
